Encapsulated functions. Built basic fastapi script to hnadle http requests to add to tracked supabase database.
